+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [v6.7.0](...)
 - Added feature custom cookie header
 - Changed `getTelemetryConfig` is now using builder.
+- Bugfix `NullPointerException` when using `ConsoleLogger`.
 ## [v6.6.0](https://github.com/PerimeterX/perimeterx-java-sdk/compare/v6.6.0...HEAD) (2023-27-04)
 - Updating readme with `customIsSensitve`, `customParametersExtraction`
 - Added an option to configure logger without slf4j using `PXConfiguration.setPxLoggerSeverity(<loggerSeverity>)`