+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Change Log
+## [v6.7.1-rc](https://github.com/PerimeterX/perimeterx-java-sdk/compare/v6.7.1-rc...HEAD)
+- extended logs
+- async activities are executed via ExecutorService
+
 ## [v6.7.0](https://github.com/PerimeterX/perimeterx-java-sdk/compare/v6.7.0...HEAD) (2023-11-05)
 - Added feature custom cookie header
 - Changed `getTelemetryConfig` is now using builder.